requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-dotenv==1.0.0
requests==2.31.0
openai==1.10.0
chromadb==0.4.22
sentence-transformers==2.3.1
scikit-learn==1.4.0
pandas==2.1.4
jinja2==3.1.3
